Pull bundler stuff into a class and extend it with OptimisticBundler
