+++ lib/bundler-utils.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const bundleJS = require('@asset-pipe/js-reader');
+const bundleCSS = require('@asset-pipe/css-reader');
+
+module.exports.bundleJS = bundleJS;
+module.exports.bundleCSS = bundleCSS;
