+++ lib/main.js
@@ -10,7 +10,7 @@ const { Transform } = require('readable-stream');
 const params = require('./params');
 const MetaStorage = require('./meta-storage');
 const Metrics = require('@metrics/client');
-const initUtils = require('./utils');
+const { booleanWithDefault } = require('./utils');
 const OptimisticBundler = require('../lib/optimistic-bundler');
 
 const DEFAULT_NODE_ENV = 'development';
@@ -110,19 +110,6 @@ module.exports = class Router extends EventEmitter {
     constructor(sink, { bundleInProcess, ...options } = {}) {
         super();
 
-        const {
-            parseBody,
-            validateFeeds,
-            bundleAndUpload,
-            endWorkers,
-            booleanWithDefault,
-        } = initUtils({ bundleInProcess });
-
-        this.parseBody = parseBody;
-        this.validateFeeds = validateFeeds;
-        this.bundleAndUpload = bundleAndUpload;
-        this.endWorkers = endWorkers;
-
         const { NODE_ENV } = process.env;
         this.options = {
             logger: emitLogger(this),
@@ -137,6 +124,7 @@ module.exports = class Router extends EventEmitter {
         this.metrics = new Metrics();
 
         this.bundler = new OptimisticBundler({
+            bundleInProcess,
             env: this.options.env,
             sink: this.sink,
             logger: this.options.logger,
@@ -204,7 +192,7 @@ module.exports = class Router extends EventEmitter {
 
         this.app.post('/publish-instructions', async (req, res, next) => {
             try {
-                const payload = await parseBody(req, res);
+                const payload = await this.bundler.parseBody(req, res);
                 await this.bundler.publishInstructions(payload, {
                     minify: booleanWithDefault(
                         req.query.minify,
@@ -318,7 +306,7 @@ module.exports = class Router extends EventEmitter {
 
             try {
                 this.emit('info', `Parsing raw feed data from body`);
-                const payload = await this.parseBody(req, res);
+                const payload = await this.bundler.parseBody(req, res);
                 this.emit(
                     'info',
                     `Successfully parsed feed data from request body. Result: ${JSON.stringify(
@@ -327,13 +315,13 @@ module.exports = class Router extends EventEmitter {
                 );
 
                 this.emit('info', `Validating parsed feed data against schema`);
-                const feedIds = this.validateFeeds(payload);
+                const feedIds = this.bundler.validateFeeds(payload);
 
                 this.emit(
                     'info',
                     `Producing and saving asset bundle for requested feeds`
                 );
-                const response = await this.bundleAndUpload({
+                const response = await this.bundler.bundleAndUpload({
                     sink: this.sink,
                     feedIds,
                     uri: this.buildUri('bundle', req.headers.host, req.secure),
@@ -468,6 +456,6 @@ module.exports = class Router extends EventEmitter {
 
     /* istanbul ignore next: invoking this method in the test has repercussions */
     async cleanup() {
-        this.endWorkers();
+        this.bundler.endWorkers();
     }
 };
