+++ lib/optimistic-bundler.js
@@ -2,12 +2,12 @@
 
 const Storage = require('./storage');
 const { hashContent } = require('./hasher');
-const initUtils = require('./utils');
 const Joi = require('joi');
 const schemas = require('../lib/schemas');
 const abslog = require('abslog');
 const { hashArray } = require('@asset-pipe/common');
 const Metrics = require('@metrics/client');
+const Bundler = require('./bundler');
 
 const PERSIST_TO_STORAGE_METRIC = {
     name: 'asset_server_persist_to_storage_timer',
@@ -24,10 +24,9 @@ const EXISTS_IN_STORAGE_METRIC = {
     description: 'Time taken for a check for existence operation from storage',
 };
 
-module.exports = class OptimisticBundler {
+module.exports = class OptimisticBundler extends Bundler {
     constructor({ bundleInProcess, ...options }) {
-        const { bundleFeeds } = initUtils({ bundleInProcess });
-        this.bundleFeeds = bundleFeeds;
+        super({ bundleInProcess });
 
         const opts = {
             sourceMaps: false,
@@ -175,7 +174,7 @@ module.exports = class OptimisticBundler {
             },
         });
 
-        const content = await this.bundleFeeds(feeds, type, {
+        const content = await super.bundleFeeds(feeds, type, {
             ...this.options,
             ...this.overrides,
         });
@@ -307,7 +306,7 @@ module.exports = class OptimisticBundler {
                 `${type} fallback bundle for tag "${tag}" already exists as "${hash}.${type}" and will not be published`
             );
         } else {
-            const bundle = await this.bundleFeeds([feed], type);
+            const bundle = await super.bundleFeeds([feed], type);
             await this.setBundle(hash, type, bundle);
             this.log.info(
                 `${type} fallback bundle for tag "${tag}" published as "${hash}.${type}"`
