+++ test/main-optimistic.test.js
@@ -3,7 +3,6 @@
 const express = require('express');
 const Router = require('../lib/main');
 const supertest = require('supertest');
-const { endWorkers } = require('../lib/utils')();
 const { hashArray } = require('@asset-pipe/common');
 const { hashContent } = require('../lib/hasher');
 const Sink = require('@asset-pipe/sink-mem');
@@ -38,7 +37,6 @@ function createTestServerFor(router) {
 }
 
 beforeAll(() => jest.setTimeout(20000));
-afterAll(() => endWorkers());
 
 const cssFeed1 = [
     {
